Statement: Find all unknown Actions and unknown Prefixes in a Statement
The primary use case of this change is for faster developer feedback.
If there are two typo'd Actions in a single IAM Statement, parliament
will only catch one; the user then makes a fix and then re-runs
parliament to find the second issue. If there is a serious Exception
with the Action syntax, we should still bail out and return False.
However, UnknownAction and UnknownPrefix are recoverable and parliament
should attempt to find all of them.