Add RESOURCE_EFFECTIVELY_STAR check
This check warns users when the combined scope of all Resource
specifiers effectively spans "*". We infer this check by seeing if there
is a wildcard Resource for every service that could possibly be
manipulated by the statement Actions.

It's still straightforward to circumvent this check by adding
a partition or an AWS account ID. Another way to circumvent this check
is by "expanding" a single Resource specifier into 3 separate partition
identifiers.